xmas_cqp/requirements.txt
# ============================================================
# XMAS-CQP Requirements
# ============================================================

# Supported Python version:
#   Python >= 3.10
#
# This file is intentionally conservative and explicitly pinned
# to ensure reproducibility across platforms (especially Windows).

# ------------------------------------------------------------
# Core OpenAI SDK (Responses API)
# ------------------------------------------------------------
openai>=2.11.0,<3.0.0

# ------------------------------------------------------------
# HTTP stack (explicitly pinned for Windows + Python 3.13 stability)
# ------------------------------------------------------------
httpx>=0.27.0,<0.28.0
httpcore>=1.0.9,<2.0.0
anyio>=4.0.0,<5.0.0
h11>=0.16.0,<0.17.0
sniffio>=1.3.0,<2.0.0
certifi>=2024.0.0

# ------------------------------------------------------------
# Schema validation
# ------------------------------------------------------------
jsonschema>=4.21.0,<5.0.0

# ------------------------------------------------------------
# Utilities (optional but recommended)
# ------------------------------------------------------------
# pydantic is reserved for future typed IR / schema models
pydantic>=2.6.0,<3.0.0

# tqdm is optional (used for progress bars in large experiments)
tqdm>=4.66.0

# ------------------------------------------------------------
# Logging / CLI
# ------------------------------------------------------------
colorama>=0.4.6; platform_system == "Windows"
pyyaml>=6.0,<7.0
